Avoid integer overflow in MillisToTicks
The overflow caused the automatic demo to trigger at every instant.
This happened after the system had been running for about 17 days.

+++ Headers/Macintosh.h
@@ -5,10 +5,10 @@
 #include "WinAPI.h"
 
 // Convert Macintosh ticks (1/60 second units) to milliseconds
-#define TicksToMillis(ticks) ((ticks) * 50 / 3)
+#define TicksToMillis(ticks) (UInt32)((unsigned long long)(ticks) * 50 / 3)
 
 // Convert milliseconds to Macintosh ticks (1/60 second units)
-#define MillisToTicks(millis) ((millis) * 3 / 50)
+#define MillisToTicks(millis) (UInt32)((unsigned long long)(millis) * 3 / 50)
 
 #define QDBlack		QDColors[0]
 #define QDYellow	QDColors[1]
